requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
tensorflow==2.10.0
h5py==3.1.0
numpy==1.21.0
Pillow==8.3.1

